- added dynamic input fields for product information to product creation panel - product informations are now displayed dynamically - user can now add an opinion

+++ Sklep Internetowy/Models/Product.cs
@@ -33,7 +33,7 @@ namespace Sklep_Internetowy.Models
         public double GetRating()
         {
             return (Rating.Count() == 0) ? 0 
-                : Rating.Sum(r => r.Rating) / Rating.Count();
+                : (double)Rating.Sum(r => r.Rating) / Rating.Count();
         }
 
     }
